MB-49512: Avoid deadlock in cancel_can_schedule test
The test schedules a task during a cancel and then checks that the
newly scheduled task runs by waiting on a condition variable until
the task runs. This is not particularly robust as it's possible for
the task to run before we wait on the condition variable and according
to the below we would not be woken again by the notify_one() call.

https://en.cppreference.com/w/cpp/thread/condition_variable/wait

```
The effects of notify_one()/notify_all() and each of the three atomic parts of wait()/wait_for()/wait_until() (unlock+wait, wakeup, and lock) take place in a single total order that can be viewed as modification order of an atomic variable: the order is specific to this individual condition variable. This makes it impossible for notify_one() to, for example, be delayed and unblock a thread that started waiting just after the call to notify_one() was made.
```

Make the test more robust by using our ThreadGate instead which handles
thread execution in either order.

Change-Id: I4131a3470790afa7b8d622d6f5677f9e032fe29b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170645
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>