MB-50941: Revert "MB-49512: Obey concurrent compaction limit when rescheduling"
Commit 280622f2 exposes and magnifies existing flaws in how we
schedule Compaction tasks - we can end up sleeping an already
scheduled Compaction task forever (when trying to limit
concurrency). This results in compaction never running for the
affected vBuckets, which is particulary problematic when cleaning up
dropped collections.

Reverting this patch in the short-term to avoid having compaction
never finish; as that is worse than exceeding the concurrent
compaction limit.

This reverts commit 280622f2947fa4c5156f2c2e78e993ea00864848.

Change-Id: I53d47880f9bd79634d14be4f4f816a3835f34652
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170741
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>