MB-49512: FollyExecutorPool - allow scheduler thread re-entrancy [1/2]
Currently the FollyExecutorPool uses a scheduler thread to deal with
any task scheduling, waking, or cancelling, and all work is done on
the CPUThreadPoolExecutor for some pool type (e.g. Writer/Reader).
One task currently run on the cpuPools is the cleaning up of task
pointers when tasks are cancelled. This is problematic during a bucket
shutdown (unregisterTaskable) as we currently must wait for task
pointers to be reset before we destruct the bucket. Should we have
contention on some cpuPool then bucket shutdown is blocked by the
running tasks to clean up the task pointers.

This is part 1 of 2 to allow the scheduler thread to perform the reset
of the task pointer rather than the cpuPool threads. If a tasks dtor
were to call back into the ExecutorPool for some reason (say to
schedule some other task) and we were to be running the reset on the
scheduler thread then we need to ensure that we do not deadlock the
thread by attempting to wait for the schedulerThread to do some work
(runInEventBaseThreadAndWait). We can support this re-entrancy by
changing these calls to runImmediatelyOrRunInEventBaseThreadAndWait
calls which run the function in the current thread if the current thread
is the scheduler thread. This may increase some scheduler thread run
times but task destruction isn't particularly common and it is less
common for task destruction to re-enter the ExecutorPool.

Change-Id: I246f15ba841778a4eaa60a54ebc082afacdbe65b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170874
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>