MB-50747: Unit test reproducing negative item count
9e08c3a2 is a fix for negative item counts that was seen in
a unit test which "resurrected" a collection, I.e. create a
collection which was previously dropped. That fix doesn't
just apply to resurrection, it applies to the case when a
delete (of an existing item) is flushed in the same flush
batch as the drop of the collection.

This commit adds a unit test for that case.

Change-Id: Ia94755052e47c089ac67706a83f24be0c27c64a1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170892
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>