MB-51011: Fix intermittent magma rollback test failure
The test tries to verify that a rollback during compaction does not
lead to an underflowed count (see MB-47566).

However, rollback may start and finish before compaction starts. The
compaction will then be cancelled early, and the result of `compactDB`
will be false.

If compaction finishes before the vb is reset, it will _not_ be
cancelled early, and `compactDB` will return true.

Ensure rollback can not occur before compaction starts, to ensure the
intended scenario is tested, and to avoid intermittent failure of
from the unpredictable result of `compactDB`.

Change-Id: I8ec82a811c08e607edcc1360c6f4a14360b38a79
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170900
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>