MB-50967: Ensure EventBase callbacks have expected order
The method:

 eventBase.runInEventBaseThread([](){...});

may be used to request a callback be executed by the event base thread.
This is used in several places to interact with FrontEndThreads.

A recent change for MB-50647 removed unnecessary queuing of
ConnHandlers for notification - now, rather than storing ptrs to
ConnHandlers in a queue for a background task, callers may immediately
call

 engine.scheduleDcpStep(cookie);

which queues a callback in the appropriate eventBase notification queue:

    connection.getThread().eventBase.runInEventBaseThread([&connection]() {
        ...
    });

This captures the associated connection by reference, and relies on the
callback being executed before the connection is freed.

However, by removing the background thread indirection, it is now
possible for FrontEndThreads to try to queue a callback for themselves.
This is typically fine, but in this case `runInEventBaseThread` may
short circuit the notification queue and execute the provided callback
immediately - it is already executing on the target thread.

By skipping the queue, callbacks may be executed in a different order
than they were logically scheduled.

This behaviour is suspected to be the origin of `scheduleDcpStep`
callbacks attempting to operate on on a destroyed connection object.

To avoid this possibility, switch to

 runInEventBaseThreadAlwaysEnqueue(...)

As the name suggests, this will always enqueue the callback.

Change-Id: Ib246b9af23341e6d2fa2f315310403bf80ad1627
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/170969
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>