MB-49469: Introduce max_checkpoints_hard_limit_multiplier
The current max_checkpoints (per vbucket) is used in the computation
for the maximum size in bytes of a single checkpoint:

  checkpoint_max_size = CMQuota / num_vbuckets / max_checkpoints

and it's also a hard limit of the number of checkpoints for the single
vbucket.

All that wants to solve 2 problems:
 1. We want to limit the size of the single checkpoint
 2. We want to limit the number of checkpoints in the single vbucket

While that apprach works fine in balanced systems, that's when the load
falls into a restricted subset of checkpoints that things start
misbehaving. Let's imagine all the load into 1 vbucket:
 - We start creating checkpoints as soon as the open one becomes full
 - At some point we hit max_checkpoints (10 by default in Neo) and we
   can't create a new one
 - Any further mutation will be queued into the 10th checkpoints, which
   can grow indefinitely at that point

Note: Just increasing max_checkpoints doesn't solve the problem as
checkpoint_max_size will be resized down accordingly and we'll just hit
the same issue.

For solving the problem, we have already worked toward the direction of
removing any max_checkpoints hard limit for the single vbucket, but that
part has been deferred to Morpheus (MB-49469).

For Neo we alleviate the problem by (i) keeping the checkpoint_max_size
computation untouched plus (ii) relaxing the max_checkpoints
constraint and allowing a higher number of checkpoints for the single
vbucket. To summarize, in Neo:
 - checkpoint_max_size = CMQuota / num_vbuckets / max_checkpoints
 - actual max checkpoints per vb = max_checkpoints * multiplier

Change-Id: I87cc74c4e573e6b643270937590827e4d15ebe79
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171008
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>