MB-49917: Add testapp tests for dcp,dcpagg,checkpoint stats
These 3 stat groups use a background task in ep-engine to gather the
data need to return the stats. Prior to the fix for MB-48816, they
accessed Cookie from a background thread without appropriate locks and
hence had data races; however we didn't detect those races via TSan in
unit tests.

This was because the only unit tests which requested those stats were
ep_testsuite ones, which do not run with the real memcached daemon and
hence didn't have the same locking usage as a real system.

Add some testapp tests (which do use the real memcached) to hopefully
catch such issues in the future.

Confirmed that this patch cherry-picked back to before the fixes for
MB-48816 (on top of commit 8ae3b4f35) does show the data races in a
TSan build.

Change-Id: I93dc90afd56fd7f86b434c0f5fa0402b39d785c2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171088
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>