MB-51512: Change ItemPager so it removes dropped collection data
If the pager encounters a key from a dropped collection it can
remove the item from memory.

This is done by linking the pageOut function with the existing dropKey
logic (after some refactoring to make the core of dropKey reusable).

This commit contains a test and changes for persistent buckets only.

Change-Id: Id509efd4ba3e3fd14696a7baf61f3a07261a2637
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171095
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>