MB-51104: Allow ItemExpel for open Checkpoints without cursors
Item expelling does not currently occur for Ephemeral vBuckets which
do not have any cursors registered - for example replica vBuckets.

This seems accidental - it is because of pre-flight check in
CheckpointManager::extractItemsToExpel() which returns early if it
finds the oldest checkpoint (where expel must begin) contains zero
cursors - with the rationale that if a Checkpoint has zero cursors in
it, we could just discard the entire Checkpoint which is cheaper than
expelling individual items.

However the check is flawed - in the case of a VBucket without any
cursors at all (such as Ephemeral without any DCP producers), the Open
Checkpoint meets this critera, however that _cannot_ be deleted.

Fixing the check is easy - also check that the checkpoint is closed
before performing the early return - however to actually make Item
Expel work for Ephemeral requires a few additional changes, given the
lowest cursor is normally used to determine where to start expelling
from.

Add a new variant of CheckpointRemoverTest memory expel tests which
checks item expel behaviour when there is zero cursors registered (as
is the case for Ephemeral buckets without any DCP producer cursors).

Change-Id: I7392781f36f73e22cf1f983befad6c8b17d25405
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171151
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>