[BP] MB-51117: Set the OSO range to only include the requested collection
Prior to this patch the range for say collection:8 was set as:

start="8"
end="9"

The kvstore by-id scan uses an inclusive range, so would of returned
a key that exactly matches the end.

The correct range is:

start="8"
end="8\xFF"

which matches only keys from collection:8

Note: This is linked to the following backport in couchstore

https://review.couchbase.org/c/couchstore/+/186964

Both are required to 1) for the new code in MB-55600 and to keep
all CV test passing (couchstore backport triggered a failure).

Change-Id: I4f248cb535fc63911c38c6883be0d0fd0f78047b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/186969
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>