MB-49512: Return enum status from KVStore::compactDB
We need to deal with aborted compactions differently to failed
compactions for MagmaKVStore and the logic should be common
between both CouchKVStore and MagmaKVStore. This change makes
the return status of compactDb an enum rather than a bool so
that we can make that differentiation and use common logic in
the callers of compactDb to update things accordingly.

Change-Id: I11957b5254814323078579444a20be3b6f86b9ef
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171176
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>