MB-35297: Add a "yield/shouldYield" methods to StatusCallback
Scans could need to yield for a variety of reasons, not just because
of a buffer being full (no_mem).

This commit adds a yield/shouldYield pair of methods to the
StatusCallback to make it clearer when code wants to pause/yield,
removing explicit use of the no_mem error to yield a scan.

Note that under the covers a engine_error is still reserved for yield,
in this case it is now the more appropriate 'temporary_failure' as the
status code.

Change-Id: If2260b3ed475d1441e8e8db1677e6809f9390f48
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171792
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>