MB-50988: Make CompactTask track requested start time
Compaction can be requested with a delay (to coalesce multiple
collection drop compactions). The delay is provided relative to "now".

Previously, this delay was used to snooze the task when created.
However, if the task can not run immediately (i.e., if there are too
many CompactTasks already running), a relative delay is hard to obey.

Make CompactTask record the requested start time, computed from
now()+delay. Now, whenever the task runs, it can check if it has
reached the requested start time, and snooze if not.

This is a step towards separating compaction delay from compaction task
concurrency control.

Change-Id: Ie0a7016c4d871a3a31747728e3e0c93dad5ad5e3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171647
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>