MB-50988: Correct sense of CompactTask::isTaskDone
isTaskDone previously returned true if the task needed rescheduling, or
had waiting cookies - i.e., if the task was _not_ done.

This is just an inconsistency with the name; it was still used to
correctly determine if the task needed to be rescheduled (the task
would be rescheduled if isTaskDone was _true_).

To make the returned value line up with the name, negate the returned
value and also the one usage of isTaskDone().

Change-Id: I8953db7bda2e0c56dba12b4631adc739bf00bc78
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171743
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>