MB-34239: Require vbid for cbstats' durability-monitor cmd
Previously it was possible to not specify a vBucket ID to cbstats'
durability-monitor, even though this behavior was not implemented.
This resulted in a 'NOT SUPPORTED with current bucket
type/configuration' error. Require passing a vbid, so clitool
raises 'Error: too few arguments - command expected a minimum
of 1 but was passed 0: []' instead.

Change-Id: I217f3975989c4fdbea77ffa847a657fc15546483
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171322
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>