MB-50988: Introduce LimitedConcurrencyTask
Some tasks need to be able to limit how many instances run
concurrently, like CompactTask.

Add LimitedConcurrencyTask to allow task concurrency to be limited by a
cb::AwaitableSemaphore.

If multiple tasks share a semaphore, the number of these tasks running
at a given time is limited by the capacity of the semaphore.

If no semaphore tokens are available, the task will snooze().

cb::AwaitableSemaphore maintains a queue of waiters, and will notify
"blocked" tasks when tokens become available. The task will then try to
acquire a token and run immediately.

This is not currently transparent to the task - it is required that the
task call:

  bool runInner() override {
     auto guard = acquireOrWait();
     if (!guard) {
         // could not acquire a token, queued for notification.
         // already snooze()-ed forever, just return true to
         // reschedule.
         return true;
     }
     // Do concurrency-limited work
  }

However, a future refactor could avoid this by, for example,
restructuring as a mixin

 LimitedConcurrency<SomeTask> task(semaphore, ...);

allowing any `SomeTask::run()` to be wrapped.

This would allow any task to be bounded by a semaphore if necessary,
without modification.

Change-Id: Ibf069081dcff31ebd99237273d6ac1962f12ef42
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171648
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>