Fix NotifiableTask default sleep time
Snoozing a task for INT_MAX seconds is treated as "sleep forever"
by GlobalTask.

NotifiableTask provides an overridable `getSleepTime()`, which
defaulted to

 std::numeric_limits<size_t>::max()

This value is larger than INT_MAX, but is not treated as "sleep
forever".

Instead, it is used as a duration relative to now(). This is likely to
overflow, and would not lead to the expected behaviour.

Fortunately, this is benign as no existing task relied on this sleep
time - the only subclass overrode this default.

Fix by updating the default sleep time.

Change-Id: I13b76d4c395307bcb8eb43255d5eb6d6b2571d3a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171336
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>