MB-45408: Pass single config string to STParam tests
Add support for passing a single config string to the STParam test
suites. This allows for greater flexibility in the test
parameterizations that we can create so rather than create new test
suites for certain configs (i.e. pitr compaction enabled) we can
instead create new parameterizations and run existing test suites
with those parameterizations (if appropriate).

Using a configuration.json style config string unfortunately isn't
trivially possible as GTest gets upset with ";" characters in the
test parameterizations so we have use ":" instead and replace them
before passing it into the engine set up.

While we migrate to the new parameterization type we still want to
support the old one for the sake of making this change smaller and
more incremental. As such, we pass the config string into the first
parameter of the existing STParam fixture and an empty second
parameter. When we finish moving the test parameterizations over we
can remove the second parameter and clean up the extra code.

Change-Id: I55ff23dda0b9c7d59c5e32c8e47f4b95884f8bec
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171422
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>