MB-50310: 2/2 Add a DCP stream from latest feature
A client may have the need to just start streaming from the current
server state, no need to obtain all historical data. A client could
discover vbucket high-seqno to do this with a stat or get-vb-seqnos
command, but to save on those extra round-trips provide a direct
way to stream from "now".

This is implemented similarly to the existing stream-to "now" feature
that is enabled by setting the correct flag on stream request.

When a client specifies the stream-from-latest flag (flag 0x40) the
following stream-request parameters are ignored

* start-seqno
* snap-start, snap-end

Internally the stream-request will set the parameters:

start-seqno = snap-start = snap-end = Vbucket::high-seqno

if vbucket-uuid is 0 then vbucket-uuid = Vbucket::uuid

The client's stream-request will return the failover table (so they
can then later resume if required) and the DCP stream will start
shipping data on the next change to the vbucket - i.e. the DCP
client will see "start-seqno+1" and onwards.

Change-Id: I4ac3c43a3ae307ae9f05366d6be0304689c02cc3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171781
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>