MB-35297: Distinguish KVStore::scan failure vs cancelled
KVStore::scan currently returns 3 status codes.

* scan_success
* scan_again
* scan_failed

With these 3 status codes there's no way for a scan to signal it cannot
continue for some state change (non fatal), e.g. vbucket or stream
closing. This limitation leads to an inefficiency that means a scan
keeps on scanning the index, even though it is doing so pointlessly.
For example when the vbucket goes away the callback sets status to
engine_errc::success and returns from the callback, yet KVStore::scan
will just keep invoking the callback for the next item in the index.
Ideally a better solution is to return an appropriate status and allow
that to get KVStore::scan to stop without triggering critical log or
increment failover stats.

This commit makes use of a linked couchstore change and refactors scan
so so that the following 4 codes can now be returned.

* ScanStatus::Success
    - The scan successfully scanned all items

* ScanStatus::Yield
    - The scan has been requested to yield due to a temporary issue,
      currently this means DCP is out of buffer space

* ScanStatus::Cancelled
    - The scan cannot continue, e.g. the vbucket or stream has gone away

* ScanStatus::Failed
    - The scan failed, this is used for when couchstore hits an error,
      e.g. some system call error or data integrity issue.

With these changes, the scan can brought to a halt as soon as one of
the scan callbacks detects the vbucket state change, setting an appropriate
code triggering the use of Cancelled.

Change-Id: I40bcde6c70a11fea4a1ec9fe2b5d2d828cc55c90
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171791
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>