MB-49512: Reset task ptr on scheduler thread during taskable shutdown
It was observed that a task being run whilst we call unregisterTaskable
on another thread would schedule the reset of the task pointer on the
cpuPool rather than the scheduler thread which caused long shutdowns as
we had to wait for the cpuPool to finish running long running tasks.
These long bucket shutdowns caused rebalance failures.

To fix this issue, reset the task pointer on the scheduler thread if we
are currently shutting down a taskable.

Change-Id: I3e6d13ba1004f32e960280123edd860487d566ff
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171920
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>