MB-51345: Stop ep_testsuite tests relying on dev assert behaviour
Assorted testsuite tests previously relied upon a call to:

 CheckpointMemRecoveryTask::attemptCheckpointRemoval()

being made in:

 CheckpointMemRecoveryTask::runInner()

which, under eager checkpoint removal, is only made as part of a dev
assert.

Fix the tests to correctly drive KV such that the test guards the
desired behaviour regardless of dev asserts.

A future patch should remove visible side effects of that dev assert,
or remove it entirely (it was a sanity check, which has not failed
since introduction). This patch is a test-only fix to unblock Neo CV.

Change-Id: I9bb5879339d5cd8047716d0b741cd27f38693dae
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171932
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>