Bump compaction timeout in PiTR test
The test is often failing locally as it takes more than 2 seconds to
run a compaction. Adding extra logs showed that slow flushes were
the cause. Whilst the test probably shouldn't take that long to flush,
it shouldn't cause us to fail due to a compaction command taking too
long. Bump the timeout to the duration that we run the compaction
thread for, if we don't compact enough during the test then we'll
fail a check at the end on the number of compactions run instead.

Change-Id: I9274b0093448e7aaf15d347a3de38c2dfe35981c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/171978
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>