MB-51391: Avoid crash when attempting to double-delete vBucket
There is a latent race condition in KVStore::deleteVBucket, where the
function performs an initial initial check for a null pointer (and
returns early), but we do not check again after acquiring an
exclusively locked VBucketPtr.

This means that if another thread also is processing a DEL_VBUCKET
command, then the state of the VBucket could change during the
function's execution,

This "double DEL_VBUCKET" has been observed to occur if the first
DEL_VBUCKET takes "too long" from ns_server's pov, and it issues a
second DEL_VBUCKET request.

Fix by moving the null check until after we have acquired the
LockedVBPtr.

Change-Id: I21e5e26444e1ed4386122b55d6ca893f7da33526
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/172098
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>