MB-51408: Don't miss closing the open checkpoint at memory recovery
There are a couple of DEV_ASSERTS in the CheckpointMemRecoveryTask that
wrap an important step of the checkpoint memory recovery logic: closing
the open checkpoint and creating a new one, if possible.

Recently we set DEV_ASSERTS=OFF for Neo, which caused the task missing
that step. Consequence is that we might enter unrecoverable TempOOM
states, eg:
 1. Store mutation A in vb1/checkpoint1 - no OOM
 2. Store mutation B in vb2/checkpoint1 - OOM
 3. CheckpointMemRecoveryTask runs and doesn't close the open
    checkpoints
 4. The same task can't expel anything as we never expel high-seqno
 5. Unrecoverable OOM

This patch fixes by reverting step (3) back into the code.
Covered by unit tests for both persistent/ephemeral.

Change-Id: Id51b837bd9500a26c89ae7999a8a4a60cd1930b5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/172215
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Well-Formed: Restriction Checker