MB-51413: Don't warmup vBuckets without state
It's possible for both couchstore and magma backends to create a
"vBucket" (header/KVStore respectively) without persisting the
vbucket_state. Warming up such a vBucket currently causes us to
warmup a vBucket with a defaul initialized vBucket state. Doing so
should be benign, ns_server should spot the vBucket in the incorrect
state post-warmup and correct it. If we were to have some IO error
during warmup though and we were to fail to persist the vBucket state
to disk during the PopulateVBMap phase of warmup then the subsequent
scan of the vBucket would find no vbucket_state and an exception would
be thrown. We'd already be in a bad state, having been unable to persist
something to disk, but the warmup would then be aborted and no more
items would be loaded into memory.

There is no point even trying to scan a vBucket in such a state, it
should not be possible for us to have a vBucket without a state that
actually has data, and such a vBucket is in an invalid state anyway.
Treat a vBucket in this state similary to one with a corrupt snapshot
and don't load it at all, ns_server will correct the vBucket states
post-warmup.

Change-Id: I53d31a89828659f24d98d9083c574a661a0e7f4e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174475
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>