MB-51373: Inspect and correct Item objects created by KVStore
Create a function that Couch/Magma-KVStore should call when
they have created an Item from the underlying stored data.
The function inspects the Item for datatype anomalies and if
found logs and corrects the Item preventing exceptions
occurring further up the stack.

In this case we check for an Item with no value, but a datatype,
which in the case of datatype=xattr can cause faults in xattr
inspection code.

Change-Id: I10b87297797c920a52a5a2e6a186b406850ae978
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/172444
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>