MB-50841: Add descriptions from stats.org to json definitions
stats.org holds a lot of documentation about individual stats, but
there's nothing ensuring it is up to date.

By bringing documentation into the json definitions, there is now a
single source of truth to refer to, and it is simpler to ensure it is
up to date (e.g., a removed stat won't leave a description hanging
around to confuse readers).

This patch pulls in stats with exact key matches in stats.org. Many
stats remain undocumented, but this can be improved over time.

For now, the "description" field of the stat def is purely
informational. It could _potentially_ be exposed for Prometheus as HELP
text[1], however this would increase the size of scrape responses by a
non-trivial amount, and is not currently planned.

[1]:https://github.com/prometheus/docs/blob/c6017097be275e13c628d0a5d9622e401326da3b/content/docs/instrumenting/exposition_formats.md#comments-help-text-and-type-information

Change-Id: If103ca8027e7183b334b207c25f53325462a2ffb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/172781
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>