MB-51298: Remove checkpoint re-use optimization in CheckpointManager
Removed the ability of the CheckpointManager to re-use empty
checkpoint. Which could occur when a replica vbucket received a
snapshot and the current open checkpoint had no items in it. This
could occur after a CheckpointManager had been created but nothing
had been added to the checkpoint, or a old active flipped to being
a replica but it's open checkpoint didn't have any non meta items
in it.

This will help us to make more of the checkpoint immutable
in the future patches. This will help us to better reason about the
state of a checkpoint and what are valid transitions for snapshot
ranges.

Due to us now sometimes creating empty checkpoints in
some unit test, this patch also make amendments to multiple tests.

Change-Id: Ieebe09e4e17761f5a15ec3e524a419941eab7cd2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173044
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>