MB-51606: Notify DM for persistence of dead vBuckets
We might have a vBucket transition from active->dead->replica during
a rebalance quickly enough that the vBucket is not deleted from
memory. To ensure that the DM state is correct at the end of the
transition we should notify the DM of persistence in any state. The
issue here manifested in the following way:

1) vBucket was active
2) vBucket starts sync write requiring persistence (seqno 1)
3) vBucket set to dead
4) vBucket persists sync write from step 3
5) vBucket set to replica
6) vBucket connects to new active and sends no seqno ack

If step 4 had run before step 3 then we would have moved the HPS
and at step 6 sent a seqno ack of 1. With this order of events though
the DM is not moving the HPS as the persistence notification is
ignored for dead vBuckets. At step 6 even though our HPS is logically
1, it is not acked back to the new active as the DM think it is still
0. This results in the write getting stuck until another durable write
is performed.

Correct this issue by notifying vBuckets of persistence regardless of
state.

Change-Id: I045b04f33a9466d52d1c39a9c9c9c78013946409
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174464
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>