MB-51606: Don't throw if PDM of dead vBucket tries to ack
We need to notifyPersistence to the DM of dead vBuckets to ensure
that the HPS is moved appropriately so that if the vBucket is
set to replica then it acks back a correct value to the active.
Dead vBuckets should not be allowed to ack and currently they throw
and assertion if they try to. Modify the function to ignore the
attempted ack for dead vBuckets rather than throw.

Change-Id: I69246810bee51f2a0fe6f25ae2b9ecfc23254d80
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174465
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>