Refactor CheckpointManager::getItemsForCursor() to take a ref over ptr
Refactor CheckpointManager::getItemsForCursor() to take a ref instead of
a ptr for its cursor argument, as the cursor is always required to be
not nullptr otherwise we would log a warn and return nothing. Making
it a ref makes this more explicit when reading and allows us to remove
the if statement within the method.

Change-Id: I8a865fa55c7418d3fdbda418d7eaa1f266944e2e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173083
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>