MB-35297: Continue/Cancel RangeScan on an I/O task
Add RangeScanContinueTask (READER) which is a task to be scheduled
in response to scan continue or cancel requests.

For continue, the "scanning" occurs on the I/O task.
For cancel, the destruct (close of snapshot) occurs on the (same) I/O
task.

In this commit the task "model" is intentionally simple, 1 new task
created per continue or cancel. The focus of the commit is the
scan continue/cancel logic and a future commit will look to improve
the task model by limiting the tasks and/or limiting the concurrency.

Change-Id: I96daac2c6b81da145f21754e17c519de13cafc03
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173472
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>