MB-35297: Add an optional random sampling feature to RangeScan
The range scan functionality can be adapted so that a client
could request that the scan returns a randomly chosen sample
of keys or documents (an alternative to the the resident only
GetRandomKey)

An optional parameter is added to the create path allowing
for a PRNG seed and sample size to be specified. Note that
when this mode is used, the range should be the entire ByID
range for the collection (this is not enforced in this commit).

When sampling is enabled the following happens.

1) During RangeScan construction (after opening the snapshot)
test that the collection stores enough items so that we could
return the requested sample-size.

2) Compute a probability (0.0 - 1.0) which is sample request
divided by collection size. E.g. if 500 samples are requested
and the collection is size is 1000, the probability is 0.5

3) Construct a std::bernoulli_distribution using the probability
as the 'p' parameter and also construct a PRNG.

4) As the scan loads a key, generate true/false using the
distribution, true and the key is included in the scan output.

Note this commit adds a "totalLimit" member, which when !0 limits
the number of keys a scan can return (for the entire scan). This
is only used by sampling, but is added 'generically' and could
be used by other functionality.

Change-Id: I3ca3af43e5e04d621a4f1df164fefcaebd106149
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174519
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>