MB-50984: Remove num-item-based checkpoint creation
Since Neo we have an implicit upper bound on the max number of
checkpoints (per bucket) that is logically given by (CMQuota /
CheckpointMaxSize).

Still, there are code paths that interfere with that and possibly break
that invariant. One of those is chk_max_items, which triggers checkpoint
creation when the user-defined max num of items has been queued into a
single checkpoint. Here we remove that trigger.

Change-Id: I2c49065a3d1b05493df8ce9865982d6393f9373f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173142
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>